生成Service

+++ izumiJava-demo/src/main/java/com/izumiJava/mappers/UserInfoMapper.java
@@ -5,7 +5,7 @@ import org.apache.ibatis.annotations.Param;
 /**
  * @Description:用户信息Mapper
  * @author:izumi
- * @Date:2023/05/26
+ * @Date:2023/05/27
  */
 public interface UserInfoMapper<T, P> extends BaseMapper {
 	/**
@@ -16,12 +16,12 @@ public interface UserInfoMapper<T, P> extends BaseMapper {
 	/**
 	 * 根据UserId更新
 	 */
-	Integer updateByUserId(@Param("bean") T t, @Param("userId") Integer userId);
+	Long updateByUserId(@Param("bean") T t, @Param("userId") Integer userId);
 
 	/**
 	 * 根据UserId删除
 	 */
-	Integer deleteByUserId(@Param("userId") Integer userId);
+	Long deleteByUserId(@Param("userId") Integer userId);
 
 	/**
 	 * 根据NickName查询
@@ -31,12 +31,12 @@ public interface UserInfoMapper<T, P> extends BaseMapper {
 	/**
 	 * 根据NickName更新
 	 */
-	Integer updateByNickName(@Param("bean") T t, @Param("nickName") String nickName);
+	Long updateByNickName(@Param("bean") T t, @Param("nickName") String nickName);
 
 	/**
 	 * 根据NickName删除
 	 */
-	Integer deleteByNickName(@Param("nickName") String nickName);
+	Long deleteByNickName(@Param("nickName") String nickName);
 
 	/**
 	 * 根据Email查询
@@ -46,12 +46,12 @@ public interface UserInfoMapper<T, P> extends BaseMapper {
 	/**
 	 * 根据Email更新
 	 */
-	Integer updateByEmail(@Param("bean") T t, @Param("email") String email);
+	Long updateByEmail(@Param("bean") T t, @Param("email") String email);
 
 	/**
 	 * 根据Email删除
 	 */
-	Integer deleteByEmail(@Param("email") String email);
+	Long deleteByEmail(@Param("email") String email);
 
 	/**
 	 * 根据QqOpenId查询
@@ -61,11 +61,11 @@ public interface UserInfoMapper<T, P> extends BaseMapper {
 	/**
 	 * 根据QqOpenId更新
 	 */
-	Integer updateByQqOpenId(@Param("bean") T t, @Param("qqOpenId") String qqOpenId);
+	Long updateByQqOpenId(@Param("bean") T t, @Param("qqOpenId") String qqOpenId);
 
 	/**
 	 * 根据QqOpenId删除
 	 */
-	Integer deleteByQqOpenId(@Param("qqOpenId") String qqOpenId);
+	Long deleteByQqOpenId(@Param("qqOpenId") String qqOpenId);
 
 }
